Updated makefile to install and uninstall binary script at $(BINDIR)/$(BINNAME) and to use the correct $(MODNAME) within the script.  Currently installs as osbf3 for testing purposes.
Added two new whitelist/blacklist show commands, which
can emit the command-line stuff needed to recreate all or
part of a list, or to delete all or part of a list.
Basically this sequence of a commands is something a user
could edit and then run in order to make wholesale
whitelist or blacklist changes (since the internal
representation is not very convenient).