The new
  osbf3 cache-report nr

command actually appears to work.

Now all we have to do is recognize and handle a subject line
equal to

  batch_train password

and I think we've implemented all the functionality the package needs.

(Although if you want the cache-report somewhere other than stdout,
there's still some work to do; I just did

  osbf3 -udir ~/osbf-lua cache-report nr | sendmail -t

and this worked like a charm.  This would be fine in a cron job as
well.)