Combines two major changes:
  1. Refactored the code so that most failing code calls lua_error
     (from C) or error (from Lua) rather than returning nil as the
     first result.  The primary benefit is a simplification in
     learn.lua.

     I sense there is too much overlap now among util.validate,
     util.insist, util.error, and util.die.  I'm not quite sure what
     the correct organizing principle is, but it needs work.

  2. I've added functionality for multiple classifications.  An
     earlier version was tested slightly, but this version is not.
     The multiple classifications are broken down using a binary
     search tree.  A user can specify a tree by hand (going on what
     messages are likely to be similar) or can simply rely on OSBF-Lua
     to create a tree from a list (which will minimize the number of
     classifications but which may be less good at distinguishing
     messages).

     Here's an example:

       multi = { classes = { 'spam', { 'calls', 'work', 'ham' } },
                 tags = { sure = { spam = '--' },
                          unsure = { spam = '-', calls = 'c',
                                     work = 'w', ham = '+' },
                          sfid = { spam = 's', calls = 'c',
                                   work = 'w', ham = 'h' },
                        },
                 threshold = { spam = 20, calls = 15, work = 20, ham = 10 },
               }

     Here 'spam' is a tree node by itself, whereas it is left up to
     the system how to create a binary tree of 'calls', 'work', and 'ham'.

     Under the 'multifilter' command each message gets two additional
     headers: one indicating the classification and one indicating if
     training is needed.


The base classification system assumes that there can be multiple
databases per class (spam/ham) but that there are exactly two classes,
whereas the multiclassification system assumes that there are exactly
two databases in each call to core.classify, but the number of classes
(and databases) is not fixed.  These two different sets of assumptions
make it very difficult to share code.  It would be very pleasant if a
way could be found to generalize one of these systems sufficiently to
make one a special case of the other.   One approach would be to add
another field to the 'multi' table giving names of databases for
classifications that use more than one database.  As long as the list
contains only two classfications, the multiclassification system
should then let well enough alone.

Another big problem is that we really lack a regression-test suite.
We need an automatic way to train on a bunch of messages and then test
further messages and check scores.  Just exercising all the code would
be a good idea.


Let me know what you think.


Norman