- fixed a_priori_prob in osbf_bayes.c - changed tone and tone_msg_and_reinforce_header to use class probability  instead of database probability
Classification now works fine for 2 classes but the scores are not calibrated, resulting
in poor (1 - ROCA)%. This is probably because of the change in how pR is calculated. For
some reason, the previousmethod avoided degenerated values.i
Note that now the sum of the two pRs don't add up to 0 when one of them
is 180 or -180. If we replace 180 or -180 by the negative of the other pR
(two classes) it seems we'd get better scores - didn't test yet.