Lots of work to (a) remove multiple-db-per-class code and (b) get right answers.  Better progress on (a) than on (b) unfortunately.
It's clear that calculations are exquisitely sensitive to arithmetic.
I've given up doing any subtraction and have fallen back on Fidelis's
suggestion that we simply sum up the probabilities of things that are
not in the class desired.  This seems to produce results that are
close to the original, but they start to diverge after about 13
trials.  Unfortunately the overall result is 10% worse than the
original.

: nr@homedog 10174 ; ./roc.lua result
1-ROCA%: 0.065383
: nr@homedog 10175 ; ./roc.lua oresult
1-ROCA%: 0.060771

I have no idea what's going on here---with only two classes, order of
addition shouldn't matter---but I suspect I'm not using OSBF_SMALLP in
the way it was used before.  Take a look at the new version of core.pR
(the C code); if you approve, we can remove any mention of smallP from
learn.lua.  More likely, we should try both ways and see if fixing
smallP can get exact answers as before.


Norman