OK, it reads the old databases and writes new ones.
I cleaned up the stats code; mmap should be cheap if the file is read-only.
Same applies to the increment code.

Given the context of the code, where it does not have access to other
databases, I saw no opportunity to recover good information about
false positives.  We could easily write something for the case of
two databases: read them both, ignore the false positives, and
reconstruct the false positives of each from the false negatives of
the other.

I am very strongly inclined to make a 'class' an exposed datatype in Lua.
I think this will have all sorts of performance advantages.

I may work on it during the week.


N