I've split the disk-database stuff into multiple files. Format 6 is currently the native format, but I would like format 7 to become the new native format.  It adds   - magic number to identify the database to file(1)     and to identify the endianness of the words in the file and it removes   - db_id and db_flags which are not used for anything.
It would be possible (and desirable) to create a non-native reader
that will read format 7 files of the wrong endianness, but I haven't
done that.

It would be great if you could get me a couple of CRM 114 databases
and some documentation of the formats so I can make a reader for
those.

At some point I'll want to transition the native format to format 7.


Norman