Now using shared mmap for read-only files (although I can't measure the difference).
Debugging code is more suitably guarded (at least some of it).

commands.learn_msg now takes a flag to indicate whether it should be
considered to be classifying the message.

Added to statistics the number of occupants of a class, which
is classifications - fp + fn.

Minor improvements to documentation.

Added three sample scripts:

  dir-train     -- trains from files assuming that $(basename $(dirname $i))
                   is the class of $i

  label-by-hand -- finds low-confidence messages, gets them
                   hand-labelled, and adds them to the right corpus
                   (infected with some of NR's mail-display code)

  low-prob      -- finds examples of underpopulated classes, which
                   can then be added to corpora