Added some support to try to diagnose a segfault which appears to be a bug in Lua.
Refactored the usage and state types to eliminate redundant
information.  State and usage are now orthogonal, the code is simpler,
and it is impossible for the two to become inconsistent.

Changed 'reader' to 'format' because it might be a better name.
Added some documentation to osbf_disk.h.

In osbflib.h, added some terse documentation to the bucket structure.
Now that I understand what's going on, I think the fields should be
named hash1, hash2, and count, rather than hash, key, and value.
Your thoughts?