Lots of experiments with ROC curves.
It occurs to me we might want to have commands.classify call
commands.multiclassify so that we can log every classification.
This would enable us to reconstruct ROC curves for production
servers---a useful capability, I think.

I will email you some test outputs separately.

I would really like to be able to use the theory of ROC to estimate
the probability of a misclassification that lies *outside* the
reinforcement zone.  I think if we could do this calculation it would
make a good thing to appear in a paper, plus we could then have a
sound mathematical basis on which to tune the size of the
reinforcement zone.  For example, I would like the probability that a
message is mistakenly classified as spam, outside the reinforcement
zone, to be less then 1.0e-6.  This would mean a lost message every
few years.   On the other hand, I'm happy if the same probability for
the ham class is 1.0e-2, that is, 1 in 100 messages is mistakenly
classed as ham.

Of course I have no idea how to calculate any of this, but maybe some
of my colleagues do...


Norman