Worked quite a bit on logging.  My goal is that the log be machine-readable, so everything is logged as a Lua command. The argument to the log should be a table of Lua functions; the file examples/count demonstrates a simple example.
With this kind of log we should be able to recover ROC curves from
production log entries.

I've also added a CRC32 computation as an aid to identifying messages.

At present, some places in the code log an entire message.  I'm not
sure if this is ever a good idea; I think it would be good if the size
of each log entry were bounded by a constant.

Here's an example log entry:

  (...).filter {
    sfid = "sfid-H-20071229-132910-+72.14-1@spamfilter.osbf.lua",
    train = false,
    conf = {
      [ "cold-calls" ] = -132.1032744624,
      [ "mailing-lists" ] = -158.72317109103,
      cfp = -71.071137285061,
      grants = -115.05184119627,
      ham = 72.136783469711,
      work = -179.45790705575,
      jobs = -179.45790705575,
      talks = -104.8292498234,
      spam = -89.592247496377,
    },
    date = "Sat Dec 29 13:29:10 2007",
    class = "ham",
    crc32 = 949970404,
    probs = {
       [ "cold-calls" ]  = 1.5597499917256e-225,
       [ "mailing-lists" ]  = 1.1873701586729e-270,
      cfp = 4.3386567500248e-122,
      grants = 1.2410501784673e-196,
      ham = 1,
      work = 8.5303303599672e-306,
      jobs = 8.5303303599672e-306,
      talks = 2.6315728497367e-179,
      spam = 1.7605285406791e-153,
    },
  };

The default config file currently contains two flags, 'log_incoming'
and 'log_learned', which are no longer used by any code.  I feel
pretty strongly that these options should be eliminated; if people
want a permanent record of messages, they can use a cache and not
expire it.  What do you think?

P.S. There may be some bugs lurking in some of the messages involving
*not_necessary*.