cache.try_recover has finally become cache.recover with new interface (I needed it for an external script)
Extended internals.lua so it can try to 'require' a missing submodule;
I have a personal osbf3.corpora module which is not part of the
standard distribution.  (I use it with the script mentioned above so
that every time I train a message it gets saved in the appropriate
corpus.)

Fixed a bug in cfg.after_loading_do: not running the function if the
config file had already been loaded.

Fixed a bug in util: not requiring log correctly.