I've fixed the bug; it had to do with bflags not being initialized properly on repeat usages of a single class.
The whole bflags thing has been bugging me for a while, because it's
not really part of the class but is rather a temporary data structure
for the use of the classifier and the microgroomer.  I've therefore
moved its initialization to the start of the following functions:
  classify
  train
  import

Results seem to have improved:

: nr@homedog 2814 ; ./trec2 -buckets small -max 5000 tr*ll -o /tmp/r5000
Using 94321 buckets, 5000 classifications (250.0/s) require 788 learnings
      ham: 1-ROCA% = 0.065744
     spam: 1-ROCA% = 0.065744
Hand-Till: 1-ROCA% = 0.065744

: nr@homedog 2816 ; ./trec2 -o results/31jan-full -buckets trec -max 90000 tr>
Using 4000037 buckets, 37822 classifications (83.1/s) require 4300 learnings
      ham: 1-ROCA% = 0.057796
     spam: 1-ROCA% = 0.057796
Hand-Till: 1-ROCA% = 0.057796


As a 'belt-and-suspenders' measure I am reopening all classes at every
classification operation.  I believe this to be unnecessary, but I'm
going to wait and commit that change in a separate operation.

While I was trawling through the code looking for uses of bflags, I
noticed that some functions are used only in osbf_aux.c, so I made
them static.  Then it turns out some aren't used at all, so I removed
the unreachable code.  It can be recovered from CVS if necessary.


Norman