Added a __tostring metamethod for messages (msg.T). The output includes the CRC-32 for the message content, and to minimize size convert it to a string and then base64-encode the string, yielding 6 characters instead of 8 in hex.  I think this is good not only because it saves space but because the resulting string is not a Lua 32-bit address.
To do this I had to add 3 functions to the core :-(
Base64 encoding and decoding comes from LHF's code which is in the
public domain.  Converting unsigned to string was easier to do in C.

Also added a print command to 'make test', fixed a documentation error
in time parsing, and used a table of obsolete time zones instead of a
sequence of if statements.