I used the new bucket facility to compute Euclidean distances between vectors in high-dimensional feature space.  For each feature (pair of hash1 and hash2) I stored a point between 0 and 1 equal to the expected rate at which that feature occurs within the corpus (count divided by number of instances).  I then computed Euclidean distances between these normalized vectors, extending with zeroes as necessary so that every feature from every corpus is included.  I don't quite know what to make of the results but I enclose them here.
Extraction of the vectors is quite slow probably due to I/O traffic on
my server, plus the fact that I have fairly large databases.  Thoughts
about visualizing these data are welcome.

I'd quite like to see output on your binary spam/ham databases; the
script is examples/deltas.


N

cfp           to cold-calls    is  112.066 mils
cfp           to grants        is   15.024 mils
cfp           to ham           is    0.649 mils
cfp           to jobs          is   14.366 mils
cfp           to mailing-lists is    2.650 mils
cfp           to origin        is    1.706 mils
cfp           to spam          is    1.209 mils
cfp           to talks         is    1.937 mils
cfp           to work          is    9.378 mils
cold-calls    to cfp           is  112.066 mils
cold-calls    to grants        is  110.604 mils
cold-calls    to ham           is  109.776 mils
cold-calls    to jobs          is  105.609 mils
cold-calls    to mailing-lists is  113.536 mils
cold-calls    to origin        is  124.558 mils
cold-calls    to spam          is  120.925 mils
cold-calls    to talks         is  111.535 mils
cold-calls    to work          is  111.634 mils
grants        to cfp           is   15.024 mils
grants        to cold-calls    is  110.604 mils
grants        to ham           is   13.487 mils
grants        to jobs          is   24.623 mils
grants        to mailing-lists is   16.649 mils
grants        to origin        is   18.814 mils
grants        to spam          is   17.683 mils
grants        to talks         is   12.627 mils
grants        to work          is   21.915 mils
ham           to cfp           is    0.649 mils
ham           to cold-calls    is  109.776 mils
ham           to grants        is   13.487 mils
ham           to jobs          is   15.398 mils
ham           to mailing-lists is    2.146 mils
ham           to origin        is    1.666 mils
ham           to spam          is    1.105 mils
ham           to talks         is    0.947 mils
ham           to work          is    9.009 mils
jobs          to cfp           is   14.366 mils
jobs          to cold-calls    is  105.609 mils
jobs          to grants        is   24.623 mils
jobs          to ham           is   15.398 mils
jobs          to mailing-lists is   18.992 mils
jobs          to origin        is   21.806 mils
jobs          to spam          is   20.275 mils
jobs          to talks         is   16.646 mils
jobs          to work          is   22.775 mils
mailing-lists to cfp           is    2.650 mils
mailing-lists to cold-calls    is  113.536 mils
mailing-lists to grants        is   16.649 mils
mailing-lists to ham           is    2.146 mils
mailing-lists to jobs          is   18.992 mils
mailing-lists to origin        is    2.986 mils
mailing-lists to spam          is    2.536 mils
mailing-lists to talks         is    3.242 mils
mailing-lists to work          is   10.946 mils
origin        to cfp           is    1.706 mils
origin        to cold-calls    is  124.558 mils
origin        to grants        is   18.814 mils
origin        to ham           is    1.666 mils
origin        to jobs          is   21.806 mils
origin        to mailing-lists is    2.986 mils
origin        to spam          is    0.080 mils
origin        to talks         is    2.937 mils
origin        to work          is   11.527 mils
spam          to cfp           is    1.209 mils
spam          to cold-calls    is  120.925 mils
spam          to grants        is   17.683 mils
spam          to ham           is    1.105 mils
spam          to jobs          is   20.275 mils
spam          to mailing-lists is    2.536 mils
spam          to origin        is    0.080 mils
spam          to talks         is    2.422 mils
spam          to work          is   10.641 mils
talks         to cfp           is    1.937 mils
talks         to cold-calls    is  111.535 mils
talks         to grants        is   12.627 mils
talks         to ham           is    0.947 mils
talks         to jobs          is   16.646 mils
talks         to mailing-lists is    3.242 mils
talks         to origin        is    2.937 mils
talks         to spam          is    2.422 mils
talks         to work          is   10.508 mils
work          to cfp           is    9.378 mils
work          to cold-calls    is  111.634 mils
work          to grants        is   21.915 mils
work          to ham           is    9.009 mils
work          to jobs          is   22.775 mils
work          to mailing-lists is   10.946 mils
work          to origin        is   11.527 mils
work          to spam          is   10.641 mils
work          to talks         is   10.508 mils