Added two new files:
  classifier.lua -- generalizes the idea of classification and reinforcement
  mlearn.lua     -- learning with multiple classifiers

None of the code is tested but it's meant to replace much of what's in
learn.lua.  learn.lua will keep stuff to do with the cache and with
classification counts; by separating out that stuff, the mlearn is
signicantly simpler.

I've not run any of this code; I'd be grateful if you'd take a look
and see what you think.

There are two main new ideas:

  * A class may have subclasses, and we use a 'subclassifier' to
    distinguish among subclasses of a class

  * In the implementation, all the training functions are now nested
    within 'learn_msg' and 'unlearn_msg'.  This change seems to
    simplify the code significantly by reducing the number of
    parameters and results passed around.   But there's a risk of
    getting the best class or confidence table wrong...


Norman